+++ island-app/src/main/java/module-info.java
@@ -14,6 +14,7 @@ module com.island.app {
     requires spring.boot.autoconfigure;
     requires spring.context;
     requires spring.web;
+    requires spring.webmvc;
     requires spring.beans;
     requires spring.core;
     requires spring.messaging;